Workers/tests/approved/semantics/reporting-errors/001.html

<!--
var port;
onerror = function(a,b,c,d) {
  // will return undefined, thus the error is "not handled"
  // so error should be reported to the user, but this test doesn't check
  // that.
  // just make sure that this method is invoked with four arguments
  var log = '';
  if (arguments.length != 4)
    log += 'got ' + arguments.length + ' arguments, expected 4. ';
  if (typeof a != 'string')
    log += 'first argument wasn\'t a string. ';
  if (b != location.href)
    log += 'second argument was ' + b + ', expected ' + location.href + '. ';
  if (typeof c != 'number')
    log += 'third argument wasn\'t a number. ';
  if (typeof d != 'number')
    log += 'fourth argument wasn\'t a number. ';
  port.postMessage(log);
}
onconnect = function (e) {
  port = e.ports[0];
  y(); // will "report the error"
}


/*
-->
<!doctype html>
<title>shared worker, not handled</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
async_test(function() {
  var worker = new SharedWorker('#', '');
  worker.port.onmessage = this.step_func(function(e) {
    assert_equals(e.data, '');
    this.done();
  });
});
</script>
<!--
*/
//-->
